llm/requirements.txt
langchain
langchain-community
supabase
python-dotenv
fastapi
uvicorn
pydantic
langchain_ollama